nilfs2_ss_manager: use tmpfs if mount point is on nilfs
If the mount point of snapshots exists on nilfs, creating a snapshot
triggers another snapshot since a directory will be created there for
the previous snapshot.  Thus, the snapshot creation process will fell
into an infinite loop.

This prevents the chain reaction by inserting tmpfs on the mount point
if it's on nilfs.

Signed-off-by: Ryusuke Konishi <ryusuke@osrg.net>
Signed-off-by: Jiro SEKIBA <jir@unicus.jp>