nilfs2.py: select latest checkpoint rather than prior checkpoint
The lscp method of nilfs2.py shrinks checkpoints having the same
timestamp, but current implementation selects the first checkpoint if
continuing checkpoints have the same timestamp.

This will change the policy so that it leaves the last checkpoint or a
snapshot (if exists) among cointantaneous checkpoints.

The reason is, in general, preceding checkpoints are transitive and
the later ones are more important.

Signed-off-by: Ryusuke Konishi <ryusuke@osrg.net>
Signed-off-by: Jiro SEKIBA <jir@unicus.jp>